requirements.txt
streamlit
tensorflow
opencv-python-headless
numpy
pillow
pandas
matplotlib
